requirements.txt
Flask
flask-cors
google-generativeai
python-dotenv
colorama
google-auth